Factor out list generation.
